Fix remaining compiler warnings: sign comparison, VLA, and search path
- Fix sign comparison warning in soundmacos.cpp by casting to size_t
- Suppress VLA (variable length arrays) warnings with -Wno-vla-cxx-extension
  VLAs are used in rs.cpp for Reed-Solomon error correction and are a
  common Clang extension that works correctly
- Remove invalid pkgconfig path from QMAKE_LIBDIR (pkgconfig is not a lib dir)

These changes eliminate 10 more warnings.

+++ src/macos/soundmacos.cpp
@@ -394,7 +394,7 @@ private:
 		std::unique_lock<std::mutex> Lock(buffers_mutex);
 		buffers.push_front(buffer);
 		queued_size += buffer->mAudioDataByteSize;
-		if (buffers.size()+2 >= buffers_for_one_second)
+		if (buffers.size()+2 >= static_cast<size_t>(buffers_for_one_second))
 		{
 			recyle_buffer();
 			overruns++;
